+++ packages/plugin-docs-parser/CHANGELOG.md
@@ -1,3 +1,15 @@
+# v0.0.5 (Tue May 26 2026)
+
+#### 🐛 Bug Fix
+
+- Plugin Docs: Resolve relative asset paths from doc file directory [#2611](https://github.com/grafana/plugin-tools/pull/2611) ([@sunker](https://github.com/sunker))
+
+#### Authors: 1
+
+- Erik Sundell ([@sunker](https://github.com/sunker))
+
+---
+
 # v0.0.4 (Thu Mar 19 2026)
 
 #### 🐛 Bug Fix